% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parseExample}
\alias{parseExample}
\title{Simple JSON Parsing Example}
\usage{
parseExample()
}
\value{
Nothing is returned, the function is invoked for its side effect.

This example is adapted from a blogpost announcing an earlier \sQuote{simdjson}
release. It is of interest mostly for the elegance and conciseness of its C++
code rather than for any functionality exported to R.

The function takes no argument and returns nothing.
}
\description{
Simple JSON Parsing Example
}
\examples{
parseExample()
}
