\name{RcppSimdJson-package}
\alias{RcppSimdJson-package}
\alias{RcppSimdJson}
\docType{package}
\title{\packageTitle{RcppSimdJson}}
\description{\packageDescription{RcppSimdJson}}
\section{Package Content}{\packageIndices{RcppSimdJson}}
\author{\packageAuthor{RcppSimdJson}}
\section{Maintainer}{\packageMaintainer{RcppSimdJson}}
\keyword{package}
