% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateJSON.R
\name{validateJSON}
\alias{validateJSON}
\title{Validate a JSON file, fast}
\usage{
validateJSON(jsonfile)
}
\arguments{
\item{jsonfile}{A character variable with a path and filename}
}
\value{
A boolean value indicating whether the JSON content was
    parsed successfully
}
\description{
By relying on simd-parallel 'simdjson' header-only library JSON
files can be parsed very quickly.
}
\examples{
if (!RcppSimdJson:::.unsupportedArchitecture()) {
    jsonfile <- system.file("jsonexamples", "twitter.json", package="RcppSimdJson")
    validateJSON(jsonfile)
}
}
