% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/exported-utils.R
\name{is_valid_json}
\alias{is_valid_json}
\alias{is_valid_utf8}
\alias{fminify}
\alias{simdjson-utilities}
\title{simdjson Utilities}
\usage{
is_valid_json(json)

is_valid_utf8(x)

fminify(json)
}
\arguments{
\item{json}{JSON, or raw vectors representing json.}

\item{x}{Text, or raw vectors representing text.}
}
\description{
simdjson Utilities
}
